+++ project-base/src/Shopsys/ShopBundle/Model/Product/Product.php
@@ -1,0 +1,35 @@
+<?php
+
+namespace Shopsys\ShopBundle\Model\Product;
+
+use Doctrine\ORM\Mapping as ORM;
+use Shopsys\FrameworkBundle\Model\Product\Product as BaseProduct;
+use Shopsys\FrameworkBundle\Model\Product\ProductCategoryDomainFactoryInterface;
+use Shopsys\FrameworkBundle\Model\Product\ProductData as BaseProductData;
+
+/**
+ * @ORM\Table(name="products")
+ * @ORM\Entity
+ */
+class Product extends BaseProduct
+{
+    /**
+     * @param \Shopsys\ShopBundle\Model\Product\ProductData $productData
+     * @param \Shopsys\ShopBundle\Model\Product\Product[]|null $variants
+     */
+    protected function __construct(BaseProductData $productData, array $variants = null)
+    {
+        parent::__construct($productData, $variants);
+    }
+
+    /**
+     * @param \Shopsys\FrameworkBundle\Model\Product\ProductCategoryDomainFactoryInterface $productCategoryDomainFactory
+     * @param \Shopsys\ShopBundle\Model\Product\ProductData
+     */
+    public function edit(
+        ProductCategoryDomainFactoryInterface $productCategoryDomainFactory,
+        BaseProductData $productData
+    ) {
+        parent::edit($productCategoryDomainFactory, $productData);
+    }
+}