+++ project-base/src/Shopsys/ShopBundle/Model/Product/ProductData.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Shopsys\ShopBundle\Model\Product;
+
+use Shopsys\FrameworkBundle\Model\Product\ProductData as BaseProductData;
+
+class ProductData extends BaseProductData
+{
+}