+++ project-base/src/Shopsys/ShopBundle/Model/Product/ProductDataFactory.php
@@ -1,0 +1,33 @@
+<?php
+
+namespace Shopsys\ShopBundle\Model\Product;
+
+use Shopsys\FrameworkBundle\Model\Product\Product as BaseProduct;
+use Shopsys\FrameworkBundle\Model\Product\ProductData as BaseProductData;
+use Shopsys\FrameworkBundle\Model\Product\ProductDataFactory as BaseProductDataFactory;
+
+class ProductDataFactory extends BaseProductDataFactory
+{
+    /**
+     * @return \Shopsys\ShopBundle\Model\Product\ProductData
+     */
+    public function create(): BaseProductData
+    {
+        $productData = new ProductData();
+        $this->fillNew($productData);
+
+        return $productData;
+    }
+
+    /**
+     * @param \Shopsys\ShopBundle\Model\Product\Product $product
+     * @return \Shopsys\ShopBundle\Model\Product\ProductData
+     */
+    public function createFromProduct(BaseProduct $product): BaseProductData
+    {
+        $productData = new ProductData();
+        $this->fillFromProduct($productData, $product);
+
+        return $productData;
+    }
+}