+++ CHANGELOG.md
@@ -41,6 +41,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - Category is extended
     - Administrator is extended
     - Transport is extended
+    - Product is extended
 
 ### [shopsys/shopsys]
 #### Added